requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
accelerate=0.21.0=pypi_0
aiofiles=23.2.1=pypi_0
ale-py=0.8.1=pypi_0
altair=5.0.1=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asgiref=3.7.2=pypi_0
asttokens=2.4.0=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.1.0=pypi_0
autorom=0.4.2=pypi_0
autorom-accept-rom-license=0.6.1=pypi_0
babel=2.13.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bitsandbytes=0.41.0=pypi_0
blinker=1.6.2=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
cloudpickle=2.2.1=pypi_0
cmake=3.27.2=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.1.4=pypi_0
contourpy=1.1.0=pypi_0
cycler=0.11.0=pypi_0
daemoniker=0.2.3=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
deepspeed=0.9.5=pypi_0
defusedxml=0.7.1=pypi_0
django=4.2.6=pypi_0
django-cors-headers=4.3.0=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
envpool=0.8.3=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
farama-notifications=0.0.4=pypi_0
fastapi=0.102.0=pypi_0
fastjsonschema=2.18.1=pypi_0
ffmpy=0.3.1=pypi_0
flask=3.0.0=pypi_0
flask-cors=4.0.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.42.1=pypi_0
fqdn=1.5.1=pypi_0
fschat=0.2.24=pypi_0
gdown=4.7.1=pypi_0
gitdb=4.0.10=pypi_0
gitdb2=4.0.2=pypi_0
gitpython=3.1.32=pypi_0
glcontext=2.4.0=pypi_0
glfw=1.12.0=pypi_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gradio=3.35.2=pypi_0
gradio-client=0.2.9=pypi_0
grpcio=1.59.0=pypi_0
gym=0.21.0=pypi_0
gym-cards=0.0.1=dev_0
gym-notices=0.0.8=pypi_0
gym3=0.3.3=pypi_0
gymnasium=0.29.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
hjson=3.1.0=pypi_0
httpcore=0.17.3=pypi_0
httpx=0.24.0=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=pypi_0
imageio-ffmpeg=0.3.0=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.1.0=pypi_0
iniconfig=2.0.0=pypi_0
install=1.3.5=pypi_0
ipykernel=6.25.2=pypi_0
ipython=8.15.0=pypi_0
isoduration=20.11.0=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
json5=0.9.14=pypi_0
jsonlines=4.0.0=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter-client=8.4.0=pypi_0
jupyter-core=5.4.0=pypi_0
jupyter-events=0.7.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.7.3=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.7=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
linkify-it-py=2.0.2=pypi_0
lit=16.0.6=pypi_0
llava=1.1.1=pypi_0
llvmlite=0.41.0=pypi_0
lxml=4.9.3=pypi_0
markdown=3.5=pypi_0
markdown-it-py=2.2.0=pypi_0
markdown2=2.4.10=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
minedojo=0.1=pypi_0
mistune=3.0.2=pypi_0
moderngl=5.8.2=pypi_0
mpmath=1.3.0=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.9.2=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
networkx=3.1=pypi_0
nh3=0.2.14=pypi_0
ninja=1.11.1=pypi_0
notebook-shim=0.2.3=pypi_0
numba=0.58.0=pypi_0
numpy=1.25.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
onnx=1.14.1=pypi_0
onnxruntime=1.16.1=pypi_0
openai=0.28.0=pypi_0
opencv-python=4.8.1.78=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.10=h7f8727e_2
optree=0.9.2=pypi_0
orjson=3.9.5=pypi_0
overrides=7.4.0=pypi_0
packaging=23.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
peft=0.4.0=pypi_0
pettingzoo=1.24.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.2.1=py310h06a4308_0
pluggy=1.3.0=pypi_0
praw=7.7.1=pypi_0
prawcore=2.4.0=pypi_0
procgen=0.10.7=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.24.2=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pypi_0
pydantic=1.10.12=pypi_0
pydub=0.25.1=pypi_0
pygame=2.5.2=pypi_0
pygments=2.16.1=pypi_0
pyparsing=3.0.9=pypi_0
pyro4=4.82=pypi_0
pysocks=1.7.1=pypi_0
pytest=7.4.2=pypi_0
python=3.10.12=h955ad1f_0
python-json-logger=2.0.7=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.5.2=pypi_0
rlcard=1.2.0=pypi_0
rpds-py=0.9.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.3=pypi_0
salesforce-lavis=1.0.2=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.11.2=pypi_0
seaborn=0.13.0=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.29.2=pypi_0
serpent=1.41=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.2.2=pypi_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
stack-data=0.6.2=pypi_0
starlette=0.27.0=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.12=pypi_0
tensorboard=2.14.1=pypi_0
tensorboard-data-server=0.7.1=pypi_0
termcolor=2.3.0=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tianshou=0.5.1=pypi_0
tiktoken=0.4.0=pypi_0
timm=0.6.13=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.31.0=pypi_0
triton=2.0.0=pypi_0
types-protobuf=4.24.0.2=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023c=h04d1e81_0
uc-micro-py=1.0.2=pypi_0
update-checker=0.18.0=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.0.4=pypi_0
uvicorn=0.23.2=pypi_0
wandb=0.15.8=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.4=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.0=pypi_0
wget=3.2=pypi_0
wheel=0.38.4=py310h06a4308_0
xmltodict=0.13.0=pypi_0
xz=5.4.2=h5eee18b_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h5eee18b_0
